[CIT-249] Refactor command trait for mqtt
